Implement hyper-speed smart download system with advanced caching
